+++ Trarizon.Library/Collections/Helpers/EnumerableQueries/WhereSelect.cs
@@ -1,4 +1,5 @@
-﻿using Trarizon.Library.Collections.Helpers.Utilities.Queriers;
+﻿using System.Diagnostics.CodeAnalysis;
+using Trarizon.Library.Collections.Helpers.Utilities.Queriers;
 using Trarizon.Library.Wrappers;
 
 namespace Trarizon.Library.Collections.Helpers;
@@ -15,35 +16,11 @@ partial class EnumerableQuery
         return new WhereSelectQuerier<T, TResult>(source, whereSelector);
     }
 
-    private sealed class WhereSelectQuerier<T, TResult>(IEnumerable<T> source, Func<T, Optional<TResult>> whereSelector) : SimpleEnumerationQuerier<T, TResult>(source)
+    private sealed class WhereSelectQuerier<T, TResult>(IEnumerable<T> source, Func<T, Optional<TResult>> whereSelector) : SimpleWhereSelectEnumerationQuerier<T, TResult>(source)
     {
-        public override bool MoveNext()
-        {
-            const int Iterating = MinPreservedState - 1;
-            const int End = Iterating - 1;
-
-            switch (_state) {
-                case -1:
-                    _enumerator = _source.GetEnumerator();
-                    _state = Iterating;
-                    goto default;
-                case End:
-                    return false;
-                default:
-                    if (_enumerator!.MoveNext()) {
-                        if (whereSelector(_enumerator.Current).TryGetValue(out var val)) {
-                            _current = val;
-                            return true;
-                        }
-                        else goto default; // Loop 
-                    }
-                    else {
-                        _state = End;
-                        return false;
-                    }
-            }
-        }
-
         protected override EnumerationQuerier<TResult> Clone() => new WhereSelectQuerier<T, TResult>(_source, whereSelector);
+
+        protected override bool TrySelect(T inVal, [MaybeNullWhen(false)] out TResult outVal)
+            => whereSelector(inVal).TryGetValue(out outVal);
     }
 }